test: refactor test to reflect new code location
